cleaning and finalize code and removing a weatherforecast endpoint in two projects
