+++ AuthServer/appsettings.json
@@ -5,5 +5,10 @@
       "Microsoft.AspNetCore": "Warning"
     }
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "JwtSettings": {
+    "Key": "12345678901234567890123456789012",
+    "Issuer": "https://yourapp.com",
+    "Audience": "https://yourapp.com"
+  }
 }
