+++ ProtectedApi/appsettings.json
@@ -5,5 +5,10 @@
       "Microsoft.AspNetCore": "Warning"
     }
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "JwtSettings": {
+    "Key": "7xg9l6vd2AUMFkXTHeOB4D6PciSZInPv",
+    "Issuer": "https://yourapp.com",
+    "Audience": "https://yourapp.com/users"
+  }
 }
